change type of arrays to float

+++ sum_direct.cpp
@@ -9,13 +9,13 @@
 #include "sums.h"
 
 void 
-setup(int64_t N, double A[])
+setup(int64_t N, float A[])
 {
    printf(" inside direct_sum problem_setup, N=%lld \n", N);
 }
 
-double
-sum(int64_t N, double A[])
+float
+sum(int64_t N, float A[])
 {
    printf(" inside direct_sum perform_sum, N=%lld \n", N);
